% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-att1.R
\docType{data}
\name{data_att1}
\alias{data_att1}
\title{Simulated Data for a Single Attribute Assessment}
\format{
A list frame containing 4 \code{tibble} objects:
\itemize{
\item \code{resp_profiles}: A \code{tibble} with 1000 rows and 2 columns. The first
column indicates \code{resp_id} (i.e., the respondent identification number) and
the second column indicates \code{att_1} (i.e., a binary indicator for whether the
respondent mastered the first attribute).
\item \code{q_matrix}: A \code{tibble} with 2 rows and 1 column. Each row corresponds
to an assessment item, and the column entries provide a binary indicator for
whether the item assessed the attribute.
\item \code{item_params}: A \code{tibble} with 2 rows and 3 columns. Each row
corresponds to an item. The first column indicates \code{item_id} (i.e., the item
identification number). The second column indicates \code{intercept} (i.e., the
true item intercept parameter for the item). The third column indicates
\code{att_1} (i.e., the true item main effect parameter for the item).
\item \code{data}: A \code{tibble} with 2000 rows and 3 columns. The first column
indicates \code{resp_id} (i.e., the respondent identification number). The second
column indicates \code{item_id} (i.e., the item identification number). The third
column indicates \code{score} (i.e., the dichotomously scored item response).
}
}
\usage{
data_att1
}
\description{
A list containing data from a randomly simulated
single-attribute assessment.
}
\keyword{datasets}
