% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm.m2-package.R
\docType{package}
\name{dcm.m2-package}
\alias{dcm.m2}
\alias{dcm.m2-package}
\title{dcm.m2: A package for the estimating the M2 statistic for DCMs}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Jeffrey Hoover \email{jeffrey.c.hoover@gmail.com} (\href{https://orcid.org/0000-0002-0276-0308}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})
}

}
\keyword{internal}
