% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcm2-package.R
\docType{package}
\name{dcm2-package}
\alias{dcm2}
\alias{dcm2-package}
\title{dcm2: A package for the estimating the M2 statistic for DCMs}
\description{
A collection of functions for calculating the M2 model fit statistic for diagnostic classification models as described by Liu et al. (2016) \doi{10.3102/1076998615621293}. These functions provide multiple sources of information for model fit according to the M2 statistic, including the M2 statistic, the *p* value for that M2 statistic, and the Root Mean Square Error of Approximation based on the M2 statistic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/atlas-aai/dcm2}
  \item Report bugs at \url{https://github.com/atlas-aai/dcm2/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey Hoover \email{jeffrey.c.hoover@gmail.com} (\href{https://orcid.org/0000-0002-0276-0308}{ORCID})

Authors:
\itemize{
  \item W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})
}

Other contributors:
\itemize{
  \item Wenchao Ma \email{wenchao.ma@ua.edu} (Author of Mord.cpp) [contributor]
  \item University of Kansas [copyright holder]
}

}
\keyword{internal}
