% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit-transformations.R
\name{log_odds}
\alias{log_odds}
\alias{logit}
\alias{inv_logit}
\title{Log-odds Transformation}
\usage{
logit(x)

inv_logit(x)
}
\arguments{
\item{x}{A number to be transformed.}
}
\value{
A transformed double
}
\description{
These functions implement the log-odds (or logit) transformation. This is a
common transformation for psychometric models that is used to put
probabilities on a continuous scale.
}
\examples{
logit(0.6)
logit(0.5)

inv_logit(3.5)
inv_logit(0)

}
