% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{m2_fit}
\alias{m2_fit}
\title{Model Fit M2 Calculations}
\usage{
m2_fit(
  data,
  struc_params,
  pi_matrix,
  qmatrix,
  num_item_params,
  ci = 0.9,
  link = "logit",
  model_type = c("LCDM", "GDINA", "ACDM", "LLM", "RRUM", "DINO", "DINA", "BUGDINO")
)
}
\arguments{
\item{data}{A data frame containing the raw data, where there is one row per
respondent and one column per item}

\item{struc_params}{A vector containing the structural parameters of the
estimated model}

\item{pi_matrix}{An item-by-class matrix containing the probability of a
correct response by members of each latent class                              (put into its own package - to use with measr)}

\item{qmatrix}{A data frame containing the Q-matrix}

\item{num_item_params}{A vector containing the number of estimated item
parameters for each item}

\item{ci}{The confidence interval for the RMSEA, computed from the M2}

\item{link}{A character containing the link function.}

\item{model_type}{A character containing the model type (e.g., \code{LCDM})
that was estimated.}
}
\value{
A data frame containing:
\itemize{
\item \code{m2}: The M2 statistic
\item \code{df}: Degrees of freedom for the M2 statistic
\item \code{pval}: \emph{p}-value for the M2 statistic
\item \code{rmsea}: Root mean square error of approximation
\item \code{ci_lower}: Lower end of \code{ci} interval for RMSEA
\item \code{ci_upper}: Upper end of \code{ci} interval for RMSEA
\item \code{srmsr}: Standardized root mean square residual
}
}
\description{
Estimate the M2 statistic as described by Liu et al. (2016).
}
\references{
Liu, Y., Tian, W., & Xin, T. (2016). An application of
\eqn{M_2}{M2} statistic to evaluate the fit of cognitive diagnostic
models. \emph{Journal of Educational and Behavioral Statistics, 41}, 3-26.
doi:10.3102/1076998615621293
}
\concept{model fit estimators}
