% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
No return value, called to import pipe operator from magrittr
}
\description{
See \verb{magrittr::[\\\%>\\\%][magrittr::\\\%>\\\%]} for details.
}
\keyword{internal}
