% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-m2.R
\name{as_binary}
\alias{as_binary}
\title{Make Binary Profiles}
\usage{
as_binary(x)
}
\arguments{
\item{x}{The number of attributes}
}
\value{
A \code{2 ^ x} by \code{x} matrix
}
\description{
Given a number of attributes, \code{as_binary} will create all possible binary
mastery profiles.
}
\examples{
as_binary(3)
as_binary(4)
}
