% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{calc_m2}
\alias{calc_m2}
\title{Calculate the M2}
\usage{
calc_m2(
  data,
  struc_params,
  pi_matrix,
  qmatrix,
  ci = 0.9,
  link = "logit",
  model_type = c("LCDM", "GDINA", "ACDM", "LLM", "RRUM", "DINO", "DINA", "BUGDINO")
)
}
\arguments{
\item{data}{A data frame containing the raw data, where there is one row per
respondent and one column per item}

\item{struc_params}{A vector containing the structural parameters of the
estimated model}

\item{pi_matrix}{An item-by-class matrix containing the probability of a
correct response by members of each latent class}

\item{qmatrix}{A data frame containing the Q-matrix}

\item{ci}{The confidence interval for the RMSEA, computed from the M2}

\item{link}{A character containing the link function.}

\item{model_type}{A character containing the model type (e.g., \code{LCDM})
that was estimated.}
}
\description{
Calculate the M2
}
