% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-methods.R
\name{m2_fit}
\alias{m2_fit}
\title{Model Fit M2 Calculations}
\usage{
m2_fit(model, ci = 0.9, ...)
}
\arguments{
\item{model}{An estimated diagnostic classification model.}

\item{ci}{The confidence interval for the RMSEA.}

\item{...}{Unused, for extensibility.}
}
\value{
A data frame containing:
\itemize{
\item \code{m2}: The M2 statistic
\item \code{df}: Degrees of freedom for the M2 statistic
\item \code{pval}: \emph{p}-value for the M2 statistic
\item \code{rmsea}: Root mean square error of approximation
\item \code{ci_lower}: Lower end of \code{ci} interval for RMSEA
\item \code{ci_upper}: Upper end of \code{ci} interval for RMSEA
\item \code{srmsr}: Standardized root mean square residual
}
}
\description{
Estimate the M2 statistic as described by Liu et al. (2016).
}
\references{
Liu, Y., Tian, W., & Xin, T. (2016). An application of
\eqn{M_2}{M2} statistic to evaluate the fit of cognitive diagnostic
models. \emph{Journal of Educational and Behavioral Statistics, 41}, 3-26.
\doi{10.3102/1076998615621293}
}
