% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Mord}
\alias{Mord}
\title{Calculate the Mord statistic}
\usage{
Mord(item_no, LCprob, prior)
}
\arguments{
\item{item_no}{The item number}

\item{LCprob}{Item parameters}

\item{prior}{Class membership probabilities}
}
\description{
Calculate the Mord statistic
}
\author{
Wenchao Ma, https://github.com/Wenchao-Ma/GDINA
}
